pyproject.toml
[tool.poetry]
name = "clashpool"
version = "0.1.0"
description = "For providing live clash pools with proxies inside"
authors = ["Guannan Ma"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cup = ">3.2"
tomlkit = "^0.11.6"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
requests = "^2.28.1"
flask = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
